refactor: move card types to libs/types
